make strategy_append_unique work for unhashables
strategy_append_unique does not work when determining
uniqueness of unhashable objects.

Adding some rudimentary support for the use case. repr()
was chosen as the fallback as it's theoretically supposed
to be a string which resolves to the object when executed,
which effectively makes it a string that one could use for
uniqueness.