fix: make predefined merge strategies work with primitive collections subtypes (#25)
Currently the predefined merge strategies for dicts, lists and sets are not being applied when the objects to be merged are different subtypes of those base types.

For instance, two dicts can be merged. Even a dict with an OrderedDict, or viceversa. But an OrderedDict and a defaultdict are treated as two completely different types and the conflict strategy is applied, although they could perfectly be merged as dicts, as they are both subtypes of dict and behave as such.