fix: Different hash for same dictionaries problem (#32)
When merging unhashable types, the hash is calculated based on `str(obj)`. The problem appears when the hash is calculated for two dictionaries like
`a = {'foo': 1, 'bar': 2}` and `b = {'bar': 2, 'foo': 1}.`

In this case, the calculated hashes obviously don't match, though from any other perspective, the dictionaries have no difference.